Add `go perft` command, and add perft tests to test suite
Perft tests search every move from a given position down to a certain
depth, and count how many leaves there are on the resulting tree.
These are a very robust way to check move generation, so this includes
some tests from the start position and from kiwipete, a position
designed to stress test chess engine move generation. This commit also
includes a UCI command to generate perft reports, which are very
helpful for debugging move generation, as well as testing performance.

I've left some of the higher depth searches commented in the test
suite, because at the moment the move generation is slow enough that
the tests time out before the searches complete.