Forbid castling from check, and check the correct squares
Check only the squares which the king moves through when we're looking
at whether we're trying to castle through check: for example, if white
is castling queenside, b1 needs to be empty for castling to be legal,
but it doesn't matter whether it's attacked.

Also, check whether the king is currently in check before castling; we
were forgetting that before.